rerouted the root of the app to the landing component - deleted Home component in app.js
