+++ src/components/Landing/Landing.js
@@ -1,26 +1,26 @@
 import React from 'react';
-import { withRouter } from 'react-router-dom';
+import Auth from 'Auth';
 
 const Landing = () => {
-  const handleGetStarted = e => {
-    e.preventDefault();
-    console.log("it's working");
+  const logOut = () => {
+    localStorage.removeItem('token');
+    localStorage.removeItem('userID');
   };
-
   return (
-    <div>
-      <header>
-        <h1>RXid</h1>
-        <p>
-          Identify your medications based on their appearance and log each
-          medication and your experiences with them
-        </p>
-      </header>
-      <section>
-        <button onClick={handleGetStarted}>Get Started</button>
-      </section>
+    <div className='homeScreen'>
+      <h2>Home screen</h2>
+      <button
+        onClick={event => {
+          event.preventDefault();
+          Auth.lock.show();
+        }}
+      >
+        Login or Register
+      </button>
+      {/* Logout needed for Auth0 testing, needs to be moved for production */}
+      <button onClick={logOut}>Logout</button>
     </div>
   );
 };
 
-export default withRouter(Landing);
+export default Landing;
